Update extension settings modal title with version number
The code changes in this commit update the title of the extension settings modal to include the version number. This provides users with a clear indication of the version they are using. The `VERSION` variable from the `globals` module is now used to dynamically display the version number in the modal title.